+++ packages/apollos-ui-connected/src/ContentNodeConnected/ContentNodeConnected.js
@@ -1,15 +1,21 @@
 import React from 'react';
 import { View } from 'react-native';
 import PropTypes from 'prop-types';
-import { Query } from '@apollo/client/react/components';
+import { useQuery } from '@apollo/client';
 import HTMLView from '@apollosproject/ui-htmlview';
 import {
+  BackgroundView,
+  BackgroundImageBlur,
   ErrorCard,
   PaddedView,
-  H2,
-  GradientOverlayImage,
+  ContentTitles,
+  ConnectedImage,
+  styled,
   named,
 } from '@apollosproject/ui-kit';
+
+import { useLike } from '../LikeButtonConnected';
+import { useShare } from '../ShareButtonConnected';
 
 import safeOpenUrl from '../safeOpenUrl';
 import GET_CONTENT_ITEM_CONTENT from './getContentNode';
@@ -20,14 +26,32 @@ const ComponentPropType = PropTypes.oneOfType([
   PropTypes.object, // type check for React fragments
 ]);
 
-const DefaultHeader = ({ isLoading, node }) => (
-  <H2 padded isLoading={isLoading}>
-    {node?.title}
-  </H2>
-);
+const NoImageSpacer = styled(({ theme }) => ({
+  height: theme.sizing.baseUnit * 4,
+}))(View);
+
+const DefaultHeader = ({ node, isLoading }) => {
+  const [isLiked, like] = useLike(node?.id);
+  const share = useShare(node?.id);
+  return (
+    <ContentTitles
+      title={node?.title}
+      summary={node?.summary}
+      featured
+      isLoading={isLoading}
+      isLiked={isLiked}
+      onPressLike={like}
+      onPressShare={share}
+    />
+  );
+};
 
 DefaultHeader.propTypes = {
-  node: PropTypes.shape({ title: PropTypes.string }),
+  node: PropTypes.shape({
+    id: PropTypes.string,
+    title: PropTypes.string,
+    summary: PropTypes.string,
+  }),
   isLoading: PropTypes.bool,
 };
 
@@ -50,44 +74,45 @@ const ContentNodeConnected = ({
   onPressAnchor,
   ImageWrapperComponent,
 }) => {
+  const { data: { node } = {}, loading, error } = useQuery(
+    GET_CONTENT_ITEM_CONTENT,
+    {
+      variables: { nodeId },
+      fetchPolicy: 'cache-first',
+    }
+  );
   if (!nodeId) return <HTMLView isLoading />;
+
+  if (!node?.htmlContent && error) return <ErrorCard error={error} />;
+
+  const coverImageSources = node?.coverImage?.sources || [];
   return (
-    <Query
-      query={GET_CONTENT_ITEM_CONTENT}
-      variables={{ nodeId }}
-      fetchPolicy={'cache-and-network'}
-    >
-      {({ data: { node } = {}, loading, error }) => {
-        if (!node?.htmlContent && error) return <ErrorCard error={error} />;
-
-        const coverImageSources = node?.coverImage?.sources || [];
-        return (
-          <>
-            {coverImageSources.length || loading ? (
-              <ImageWrapperComponent>
-                <GradientOverlayImage
-                  isLoading={!coverImageSources.length && loading}
-                  source={coverImageSources}
-                />
-              </ImageWrapperComponent>
-            ) : null}
-
-            {/* fixes text/navigation spacing by adding vertical padding if we dont have an image */}
-            <PaddedView vertical={!coverImageSources.length}>
-              <HeaderComponent
-                isLoading={!node?.title && loading}
-                node={node}
-              />
-              <HtmlComponent
-                isLoading={!node?.htmlContent && loading}
-                onPressAnchor={onPressAnchor}
-                node={node}
-              />
-            </PaddedView>
-          </>
-        );
-      }}
-    </Query>
+    <>
+      <View>
+        <BackgroundImageBlur source={coverImageSources} />
+        {coverImageSources.length || loading ? (
+          <ImageWrapperComponent>
+            <ConnectedImage
+              maintainAspectRatio
+              isLoading={!coverImageSources.length && loading}
+              source={coverImageSources}
+            />
+          </ImageWrapperComponent>
+        ) : (
+          <NoImageSpacer />
+        )}
+        <HeaderComponent isLoading={!node?.title && loading} node={node} />
+      </View>
+      <BackgroundView flex={false}>
+        <PaddedView>
+          <HtmlComponent
+            isLoading={!node?.htmlContent && loading}
+            onPressAnchor={onPressAnchor}
+            node={node}
+          />
+        </PaddedView>
+      </BackgroundView>
+    </>
   );
 };