+++ packages/apollos-ui-kit/src/DefaultCard/index.js
@@ -1,68 +1,23 @@
 import React from 'react';
-import { View } from 'react-native';
 import PropTypes from 'prop-types';
 
 import { withTheme, named } from '../theme';
-import styled from '../styled';
-import Card, { CardImage, CardLabel, CardContent } from '../Card';
-import { H3, BodyText } from '../typography';
+import Card, { CardImage, CardLabel } from '../Card';
 import Icon from '../Icon';
 import { withIsLoading } from '../isLoading';
 import { ImageSourceType } from '../ConnectedImage';
+import BackgroundImageBlur from '../BackgroundImageBlur';
 
-// We have to position `LikeIcon` in a `View` rather than `LikeIcon` directly so `LikeIcon`'s loading state is positioned correctly 💥
-const LikeIconPositioning = styled(
-  ({ theme }) => ({
-    position: 'absolute',
-    top: theme.sizing.baseUnit * 1.5,
-    right: theme.sizing.baseUnit * 1.5,
-  }),
-  'ui-kit.DefaultCard.LikeIconPositioning'
-)(View);
+import ContentTitles from '../ContentTitles';
 
-const LikeIcon = withTheme(
-  ({ theme, isLiked }) => ({
-    fill: theme.colors.white,
-    name: isLiked ? 'like-solid' : 'like',
-    size: theme.sizing.baseUnit * 1.5,
-  }),
-  'ui-kit.DefaultCard.LikeIcon'
-)(Icon);
-
-const Image = withTheme(
+const CoverImage = withTheme(
   () => ({
-    minAspectRatio: 1.2,
+    minAspectRatio: 1,
     maxAspectRatio: 1.78,
     maintainAspectRatio: true,
   }),
   'ui-kit.DefaultCard.Image'
 )(CardImage);
-
-const Content = styled(
-  ({ theme }) => ({
-    alignItems: 'flex-start', // needed to make `Label` display as an "inline" element
-    paddingHorizontal: theme.sizing.baseUnit * 1.5, // TODO: refactor CardContent to have this be the default
-    paddingBottom: theme.sizing.baseUnit * 2, // TODO: refactor CardContent to have this be the default
-  }),
-  'ui-kit.DefaultCard.Content'
-)(CardContent);
-
-const Summary = styled(
-  ({ theme }) => ({
-    paddingTop: theme.sizing.baseUnit,
-  }),
-  'ui-kit.DefaultCard.Summary'
-)(BodyText);
-
-// We have to position `renderLabel/CardLabel/LabelComponent` in a `View` so `Label`s loading state is positioned correctly 💥
-// We also only render this component if we have a label.
-const LabelPositioning = styled(
-  ({ theme }) => ({
-    marginTop: -theme.sizing.baseUnit * 2,
-    marginBottom: theme.sizing.baseUnit,
-  }),
-  'ui-kit.DefaultCard.LabelPositioning'
-)(View);
 
 const LiveIcon = withTheme(({ theme }) => ({
   name: 'live-dot',
@@ -74,20 +29,18 @@ const renderLabel = (isLoading, LabelComponent, labelText, summary, isLive) => {
   let ComponentToRender = null;
 
   if (LabelComponent) {
-    ComponentToRender = <LabelPositioning>{LabelComponent}</LabelPositioning>;
+    ComponentToRender = LabelComponent;
 
     // this always shows a loading state for labels
   } else if (labelText || isLoading || isLive) {
     ComponentToRender = (
-      <LabelPositioning>
-        <CardLabel
-          isLive={isLive}
-          hasSummary={summary}
-          title={labelText || (isLive ? 'Live' : null)}
-          type={'secondary'}
-          IconComponent={isLive ? LiveIcon : null}
-        />
-      </LabelPositioning>
+      <CardLabel
+        hasSummary={summary}
+        title={labelText || (isLive ? 'Live' : null)}
+        type={'secondary'}
+        icon={isLive ? 'live-dot' : undefined}
+        IconComponent={isLive ? LiveIcon : undefined}
+      />
     );
   }
 
@@ -96,6 +49,7 @@ const renderLabel = (isLoading, LabelComponent, labelText, summary, isLive) => {
 
 const DefaultCard = withIsLoading(
   ({
+    featured,
     coverImage,
     title,
     isLiked,
@@ -106,18 +60,23 @@ const DefaultCard = withIsLoading(
     isLive,
   }) => (
     <Card isLoading={isLoading}>
-      <Image source={coverImage} />
+      <BackgroundImageBlur source={coverImage} />
 
-      <Content>
-        {renderLabel(isLoading, LabelComponent, labelText, summary, isLive)}
-        {title ? <H3 numberOfLines={2}>{title}</H3> : null}
-        {summary ? <Summary numberOfLines={2}>{summary}</Summary> : null}
-      </Content>
-      {isLiked != null ? (
-        <LikeIconPositioning>
-          <LikeIcon isLiked={isLiked} />
-        </LikeIconPositioning>
-      ) : null}
+      <CoverImage source={coverImage} />
+      <ContentTitles
+        featured={featured}
+        title={title}
+        summary={summary}
+        isLiked={isLiked}
+        isLoading={isLoading}
+        label={renderLabel(
+          isLoading,
+          LabelComponent,
+          labelText,
+          summary,
+          isLive
+        )}
+      />
     </Card>
   )
 );
@@ -127,16 +86,13 @@ DefaultCard.propTypes = {
     PropTypes.arrayOf(ImageSourceType),
     ImageSourceType,
   ]),
-  title: PropTypes.string.isRequired,
+  featured: PropTypes.bool,
+  title: PropTypes.string,
   isLiked: PropTypes.bool,
   isLive: PropTypes.bool,
   LabelComponent: PropTypes.element,
   labelText: PropTypes.string,
   summary: PropTypes.string,
-  theme: PropTypes.shape({
-    type: PropTypes.string,
-    colors: PropTypes.shape({}),
-  }),
 };
 
 DefaultCard.displayName = 'DefaultCard';