+++ scripts/watch.js
@@ -13,17 +13,6 @@ chokidar
         console.log(`Error building ${pkg}`, err);
       } else {
         console.log(stdout, stderr);
-        console.log(`yarn lerna --scope ${pkg} exec -- yalc push`);
-        exec(
-          `./node_modules/lerna/cli.js --scope ${pkg} exec -- yalc push --changed`,
-          (err, stdout, stderr) => {
-            if (err) {
-              console.log(`Error building ${pkg}`, err);
-            } else {
-              console.log(stdout, stderr);
-            }
-          }
-        );
       }
     });
   });