+++ packages/apollos-data-connector-postgres/CHANGELOG.md
@@ -3,6 +3,81 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.25.0](https://github.com/apollosproject/apollos-apps/compare/v2.24.0...v2.25.0) (2021-06-30)
+
+
+### Features
+
+* Content Item Category model ([#2029](https://github.com/apollosproject/apollos-apps/issues/2029)) ([cb4edbf](https://github.com/apollosproject/apollos-apps/commit/cb4edbf02b6dac586f1e2438d58f9e67e261aa30))
+
+
+### Bug Fixes
+
+* Only send notifications for public journals ([#2038](https://github.com/apollosproject/apollos-apps/issues/2038)) ([e15a032](https://github.com/apollosproject/apollos-apps/commit/e15a03212ee40787f8cd7ecf23358ef6997b9a92))
+* Postgres Media bugfixes ([#2026](https://github.com/apollosproject/apollos-apps/issues/2026)) ([3d3b61d](https://github.com/apollosproject/apollos-apps/commit/3d3b61de6582750ec4609b8b30a5add3475fabf8))
+
+
+
+## [2.24.0](https://github.com/apollosproject/apollos-apps/compare/v2.23.1...v2.24.0) (2021-06-28)
+
+**Note:** Version bump only for package @apollosproject/data-connector-postgres
+
+
+
+
+
+### [2.23.1](https://github.com/apollosproject/apollos-apps/compare/v2.23.0...v2.23.1) (2021-06-24)
+
+**Note:** Version bump only for package @apollosproject/data-connector-postgres
+
+
+
+
+
+## [2.23.0](https://github.com/apollosproject/apollos-apps/compare/v2.22.0...v2.23.0) (2021-06-24)
+
+
+### Features
+
+* Content Items Connection Model ([#2017](https://github.com/apollosproject/apollos-apps/issues/2017)) ([511e3a1](https://github.com/apollosproject/apollos-apps/commit/511e3a19057e0f7955cf12976770726f4de7574a))
+* Send a push notification when adding a new comment ([#1983](https://github.com/apollosproject/apollos-apps/issues/1983)) ([f84b2b7](https://github.com/apollosproject/apollos-apps/commit/f84b2b7e50517cc5ef6c32a18f77546f41c85d9a))
+
+
+
+## [2.22.0](https://github.com/apollosproject/apollos-apps/compare/v2.21.2...v2.22.0) (2021-06-23)
+
+
+### Features
+
+* Content Item Schema ([#2006](https://github.com/apollosproject/apollos-apps/issues/2006)) ([0fc23a4](https://github.com/apollosproject/apollos-apps/commit/0fc23a42e0ea974eb7fc9a51537ea46861c2fc1d))
+* Creates media model for postgres data connector ([#2005](https://github.com/apollosproject/apollos-apps/issues/2005)) ([870aca5](https://github.com/apollosproject/apollos-apps/commit/870aca5189e92b38d89b57b6e67417a6253d3206))
+
+
+
+### [2.21.2](https://github.com/apollosproject/apollos-apps/compare/v2.21.1...v2.21.2) (2021-06-18)
+
+**Note:** Version bump only for package @apollosproject/data-connector-postgres
+
+
+
+
+
+### [2.21.1](https://github.com/apollosproject/apollos-apps/compare/v2.21.0...v2.21.1) (2021-06-17)
+
+**Note:** Version bump only for package @apollosproject/data-connector-postgres
+
+
+
+
+
+## [2.21.0](https://github.com/apollosproject/apollos-apps/compare/v2.20.0...v2.21.0) (2021-06-16)
+
+**Note:** Version bump only for package @apollosproject/data-connector-postgres
+
+
+
+
+
 ## [2.20.0](https://github.com/apollosproject/apollos-apps/compare/v2.19.0...v2.20.0) (2021-06-14)